feat (mapfilter) renames all files as map-filter
