+++ libs/ngxtension/map-filter/README.md
@@ -1,0 +1,3 @@
+# ngxtension/map-filter
+
+Secondary entry point of `ngxtension`. It can be used by importing from `ngxtension/map-filter`.
