+++ libs/ngxtension/map-filter/project.json
@@ -1,15 +1,15 @@
 {
-	"name": "ngxtension/mapfilter",
+	"name": "ngxtension/map-filter",
 	"$schema": "../../../node_modules/nx/schemas/project-schema.json",
 	"projectType": "library",
-	"sourceRoot": "libs/ngxtension/mapfilter/src",
+	"sourceRoot": "libs/ngxtension/map-filter/src",
 	"targets": {
 		"test": {
 			"executor": "@nx/jest:jest",
 			"outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
 			"options": {
 				"jestConfig": "libs/ngxtension/jest.config.ts",
-				"testPathPattern": ["mapfilter"],
+				"testPathPattern": ["map-filter"],
 				"passWithNoTests": true
 			},
 			"configurations": {
@@ -24,8 +24,8 @@
 			"outputs": ["{options.outputFile}"],
 			"options": {
 				"lintFilePatterns": [
-					"libs/ngxtension/mapfilter/**/*.ts",
-					"libs/ngxtension/mapfilter/**/*.html"
+					"libs/ngxtension/map-filter/**/*.ts",
+					"libs/ngxtension/map-filter/**/*.html"
 				]
 			}
 		}
