--- libs/ngxtension/mapfilter/README.md
@@ -1,3 +1,0 @@
-# ngxtension/mapfilter
-
-Secondary entry point of `ngxtension`. It can be used by importing from `ngxtension/mapfilter`.
